gh-139353: Add Objects/unicode_codecs_utf.c file
Rename functions:

* ascii_decode() => _PyUnicode_DecodeASCII()
* backslashreplace() => _PyUnicode_backslashreplace()
* raise_encode_exception() => _PyUnicode_RaiseEncodeException()
* unicode_decode_call_errorhandler_writer() => _PyUnicode_DecodeCallErrorHandler()
* unicode_decode_utf8() => _PyUnicode_DecodeUTF8()
* unicode_encode_call_errorhandler() => _PyUnicode_EncodeCallErrorHandler()
* unicode_encode_utf8() => _PyUnicode_EncodeUTF8()
* xmlcharrefreplace() => _PyUnicode_xmlcharrefreplace()

Move static inline functions and macros to pycore_unicodeobject.h:

* _PyUnicode_CHECK()
* _PyUnicode_UTF8()
* PyUnicode_UTF8()
* PyUnicode_SET_UTF8()
* PyUnicode_UTF8_LENGTH()
* PyUnicode_SET_UTF8_LENGTH()
