+++ PCbuild/_freeze_module.vcxproj.filters
@@ -484,6 +484,9 @@
     <ClCompile Include="..\Objects\typeobject.c">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="..\Objects\unicode_codecs_utf.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
     <ClCompile Include="..\Objects\unicode_format.c">
       <Filter>Source Files</Filter>
     </ClCompile>
