+++ PCbuild/pythoncore.vcxproj.filters
@@ -1277,6 +1277,9 @@
     <ClCompile Include="..\Objects\typeobject.c">
       <Filter>Objects</Filter>
     </ClCompile>
+    <ClCompile Include="..\Objects\unicode_codecs_utf.c">
+      <Filter>Objects</Filter>
+    </ClCompile>
     <ClCompile Include="..\Objects\unicode_format.c">
       <Filter>Objects</Filter>
     </ClCompile>
